requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
messagebird==2.0.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
